Have dev container onCreateCommand set up the venv
This adds commands to the onCreate script to create the Python
virtual environment and install the project dependencies in it.

The interpreter for the onCreate script is changed from sh to bash
because, when it was /bin/sh, this error occurs:

    .venv/bin/activate: OSTYPE: parameter not set

This strongly appears related to "set -u", which is what causes a
"parameter not set" error when parameter expansion is attempted on
an unset parameter (rather than expanding to an empty string). So
it would probably work to remove "-u", but the dev container image
is guaranteed to have bash (and to have it available as /bin/bash)
so that can be used, and the benefits of "-u" retained.