Use LF line endings on all OSes; tweak dev container settings
This sets LF (Unix-style) line endings on all systems, which is
arguably a good general practice in most projects today, but more
specifically is useful for the reason described in .gitattributes.

It reinforces this by setting it as a default in devcontainer.json
as a VS Code setting. (If VS Code's settings.json is committed to
source control, this could be moved there instead for slightly
greater effect.)

Settings for removing whitespace at the end of lines and having
a single newline at the end of a file are likewise specified for
VS Code in devcontainer.json. In contrast to the files.eol setting,
these should probably remain in devcontainer.json and not be moved
into settings.json as workspace settings, since some developers may
prefer something else here for their local configuration.