Use consistent opening docstring style
Functions in dicts.py used a different style for docstrings,
placing the (conceptual) first line immediately after the opening
triple-quotes even when the docstring is multiple lines. This style
is fine, but used nowhere else in the project. This adds newlines
to make the style consistent with that used throughout.