Rename al_sorter to sorted_al; add/expand docstrings
This renames al_sorter to sorted_al, to reflect what the function
does (and to be named in the same style as the "sorted" builtin it
uses). It adds a docstring to that function.

This also expands the docstring of adjacency, to describe the data
structures being used, and so the two functions' docstrings can be
read to recognize the important relationship between them.