Add ruff, enable various rules, fix code for them
This adds the ruff linter, and enables some rules so that it finds
PEP-8 violations and other opportunities for improvement.

- https://docs.astral.sh/ruff/
- https://pep8.org/

This also makes changes everywhere in the code where ruff
complained based on these rules, except in a small handful of cases
where it both made sense to have the rule enabled and to keep the
rule-violating code. In those cases, specific suppressions are
applied.

One change may be a bit surprising, so I mention it explicitly:
there is no "good" way to conditionally reassign a function to a
variable, when the function has to be defined anew, because:

- Assigning a lambda directly to a variable looks like it is a
  missed opportunity to use a def statement.

- Using a def statement to reassign an existing variable looks like
  an accidental name clash.

The workaround is to define a named function elsewhere and use it
by its name. In this case, that's `util.identity_function`.

In addition to the above changes, I've also changed "quickfind" to
"quick-find", as it is more commonly spelled (e.g., *Algorithms in
C++, 3rd Edition* by Robert Sedgewick, chapter 1). I think this
improves clarity because it makes it less likely for readers to
confuse quick-find and quick-union with the unrelated algorithms
quicksort and quickselect. But it is definitely not something any
ruff rules demand of us. :)