Rename source/dest as u/v when symmetric
Even with an undirected graph, some operations have a distinguished
source and destination vertex, such as during traversal. However,
in other places, there is no such distinction, so it is clearer
even to use short variable names like "u" and "v" than "source" and
"dest". This makes such a change in three places where it applies.

Another benefit of this change is that it makes some uses of the
ternary conditional operator much more compact, such that it is
readable even when written on one line. This puts those on a single
line accordingly. Those expressions are thus now written in a way
has the advantages of the forms before and after caacee4 (#12).