Remove generator exercise implementations
(And other new exercises' implementations.)

This is they can be reworked.

This also changes FIXME comments on the absent (i.e., removed)
decorator exercise implementations to "raise NotImplementedError"
for consistency.

Note that not all material that completion of the exercises would
entail adding will go in place of "raise NotImplementedError"
statements: some FIXME comments still exist, especially in the
newly added generators exercises, for filling in some missing
doctests, which in some cases are a major part of an exercise.