Add pyright CI; suppress reportFunctionMemberAccess
The reportFunctionMemberAccess suppression applies both locally and
on CI. It keeps pyright from warning about access to attributes we
add to functions. Currently this is only done in adders.py. If that
turns out to be the only place we do that intentionally, the the
suppression can be made narrower by moving it into that module.

+ Improve check for when to install setuptools on CI.